Fix according to comments
